requirements.txt
gradio>=4.0.0
google-cloud-storage>=2.10.0
google-cloud-compute>=1.14.0
python-dotenv>=1.0.0
requests>=2.31.0
